fix(dcd): extract constant and update import
