+++ src/components/[guild]/JoinModal/hooks/useJoin.tsx
@@ -1,6 +1,6 @@
 import { JoinJob } from "@guildxyz/types"
 import { CircleWavyCheck } from "@phosphor-icons/react"
-import { GUILD_PIN_MAINTENANCE } from "components/[guild]/Requirements/components/GuildCheckout/MintGuildPin/MintGuildPin"
+import { GUILD_PIN_MAINTENANCE } from "components/[guild]/Requirements/components/GuildCheckout/MintGuildPin/constants"
 import { useMintGuildPinContext } from "components/[guild]/Requirements/components/GuildCheckout/MintGuildPinContext"
 import useGuild from "components/[guild]/hooks/useGuild"
 import useUser from "components/[guild]/hooks/useUser"
