+++ src/components/[guild]/Requirements/components/GuildCheckout/MintGuildPin/MintGuildPin.tsx
@@ -26,8 +26,7 @@ import TransactionStatusModal from "../components/TransactionStatusModal"
 import GuildPinOpenseaLink from "../components/TransactionStatusModal/components/GuildPinOpenseaLink"
 import MintGuildPinButton from "../components/buttons/MintGuildPinButton"
 import SwitchNetworkButton from "../components/buttons/SwitchNetworkButton"
-
-export const GUILD_PIN_MAINTENANCE = false
+import { GUILD_PIN_MAINTENANCE } from "./constants"
 
 const DynamicActivateGuildPinModal = dynamic(
   () => import("./components/ActivateGuildPinModal")
