+++ src/components/[guild]/Requirements/components/GuildCheckout/MintGuildPin/constants.ts
@@ -1,0 +1,1 @@
+export const GUILD_PIN_MAINTENANCE = false
